requirements.txt
pyyaml==6.0.2
requests==2.30.0
pandas==2.0.1
